// frontend/src/app/models/poi.model.ts
export interface PointOfInterest {
    id?: string;
    datasetid?: string;
    recordid?: any;
    name: string;
    ville: string;
    wgs84?: number;
    description?: any;
    adresse?: any;
    telephone: string;
    longitude?: any;
    codepostal?: number;
    latitude?: any;
    titre?: any;
    categorie1?: any;
    categorie3?: any;
    categorie2?: any;

    type?: any;
    coordinates?: number;

    record_timestamp?: number;

}